+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
   "rainlang.opmeta": {
-    "deployerAddress": "0x86bE99d58c7E3d2F20F8A4aeBC792A6Ed6807548",
+    "deployerAddress": "0x446Bd29B6D2202b320718D94D13D1f121e77dfb9",
     "source": "mumbai"
   }
 }
