+++ config/config.json
@@ -5,12 +5,12 @@
             "transaction" : "0xa2aafeacc86ee8f0967481a2c89b65bfccac171d1d66c9d49b4a08013222dd63"
         },
         "expressionDeployer" :{
-            "address" : "0x0d41310d0eb1a0e1827d55c98f603263219f8a5d" ,
-            "transaction" : "0x6df9cfcd708627b053e957ffe6ce40ffad7b47b95f1bc7c16730af905fe9a068"
+            "address" : "0x446Bd29B6D2202b320718D94D13D1f121e77dfb9" ,
+            "transaction" : "0x21fe687132d1339423a9b0dac3349769fe9835c6b8e575626a78dd5688c239d1"
         } 
     },
     "hardhat" : {
         "forkBaseUrl" : "https://polygon-mumbai.g.alchemy.com/v2/" ,
-        "blockNumber" : 33806466  
+        "blockNumber" : 34436311  
     }
 }
