+++ test/decimals.ts
@@ -5,6 +5,7 @@ import { ethers  } from "hardhat";
 import { randomUint256 } from "../utils/bytes";
 import {
   eighteenZeros,
+  max_uint256,
   ONE,
   sixZeros,
 } from "../utils/constants/bigNumber";
@@ -26,7 +27,7 @@ import { getExpressionDelopyer } from "../utils/deploy/interpreter";
 import config from "../config/config.json"
 import * as dotenv from "dotenv";
 import { encodeMeta } from "../scripts/utils";
-import { prbScale, scaleRatio, takeOrder } from "../utils/orderBook";
+import { prbScale, scaleOutputMax, scaleRatio, takeOrder } from "../utils/orderBook";
 dotenv.config();
 
 
@@ -72,998 +73,6 @@ describe("Pilot", async function () {
     await deploy1820(signers[0]);
   });  
 
-
- 
-  it("should ensure only conterparties are able to takeOrders", async function () {  
-  
-    const signers = await ethers.getSigners();
-
-    const [ , alice, bob, carol] = signers;    
-
-    const aliceInputVault = ethers.BigNumber.from(randomUint256());
-    const aliceOutputVault = ethers.BigNumber.from(randomUint256());
-  
-    const aliceOrder = encodeMeta("Order_A");   
-
-    // Order_A 
-    const strategyRatio = "25e13"
-    const strategyExpression = path.resolve(
-      __dirname,
-      "../src/0-pilot.rain"
-    );
-
-    const strategyString = await fetchFile(strategyExpression); 
-
-    const stringExpression = mustache.render(strategyString, {
-      counterparty: bob.address,
-      ratio: strategyRatio
-    });   
-
-    
-
-    const { sources, constants } = await standardEvaluableConfig(stringExpression)
-
-    const EvaluableConfig_A = {
-      deployer: expressionDeployer.address,
-      sources,
-      constants,
-    }
-    const orderConfig_A= {
-      validInputs: [
-        { token: tokenA.address, decimals: 18, vaultId: aliceInputVault },
-      ],
-      validOutputs: [
-        { token: tokenB.address, decimals: 18, vaultId: aliceOutputVault },
-      ],
-      evaluableConfig: EvaluableConfig_A,
-      meta: aliceOrder,
-    }; 
-
-    const txOrder_A = await orderBook.connect(alice).addOrder(orderConfig_A); 
-
-    const {
-      order: Order_A
-    } = (await getEventArgs(
-      txOrder_A,
-      "AddOrder",
-      orderBook
-    ));
-
-     // TAKE ORDER
-
-     // DEPOSIT
-     // Deposit token equal to the size of the batch
-     const amountB = ethers.BigNumber.from("1000" + eighteenZeros);
- 
-     const depositConfigStructAlice = {
-       token: tokenB.address,
-       vaultId: aliceOutputVault,
-       amount: amountB,
-     };
- 
-     await tokenB.transfer(alice.address, amountB);
-     await tokenB
-       .connect(alice)
-       .approve(orderBook.address, depositConfigStructAlice.amount);
- 
-     // Alice deposits tokenB into her output vault
-     await orderBook
-       .connect(alice)
-       .deposit(depositConfigStructAlice);
- 
-     const takeOrderConfigStruct = {
-       order: Order_A,
-       inputIOIndex: 0,
-       outputIOIndex: 0,
-       signedContext: []
-     }; 
- 
-     let ratio = await prbScale(0,strategyRatio) 
- 
-     const takeOrdersConfigStruct = {
-       output: tokenA.address,
-       input: tokenB.address,
-       minimumInput: amountB,
-       maximumInput: amountB,
-       maximumIORatio: ratio,
-       orders: [takeOrderConfigStruct],
-     };
- 
-     const amountA = amountB.mul(ratio).div(ONE); 
- 
-     await tokenA.transfer(carol.address, amountA);
-     await tokenA.connect(carol).approve(orderBook.address, amountA); 
-
-     await assertError(
-      async () =>
-         await orderBook
-        .connect(carol)
-        .takeOrders(takeOrdersConfigStruct),
-      "",
-      "Invalid Conterparty"
-    );
-        
-  }); 
-
-  it("should ensure 24 hour delay is maintained between two batches", async function () { 
-
-    const signers = await ethers.getSigners();
-
-    const [, alice, bob] = signers;   
-
-    const aliceInputVault = ethers.BigNumber.from(randomUint256());
-    const aliceOutputVault = ethers.BigNumber.from(randomUint256());
-  
-
-    const aliceOrder = encodeMeta("Order_A"); 
-    
-    // Order_A
-    const strategyRatio = "25e13"
-    const strategyExpression = path.resolve(
-      __dirname,
-      "../src/0-pilot.rain"
-    );
-
-    const strategyString = await fetchFile(strategyExpression); 
-
-    const stringExpression = mustache.render(strategyString, {
-      counterparty: bob.address,
-      ratio: strategyRatio
-    }); 
-
-    const { sources, constants } = await standardEvaluableConfig(stringExpression)
-
-    const EvaluableConfig_A = {
-      deployer: expressionDeployer.address,
-      sources,
-      constants,
-    }
-
-    const orderConfig_A= {
-      validInputs: [
-        { token: tokenA.address, decimals: 18, vaultId: aliceInputVault },
-      ],
-      validOutputs: [
-        { token: tokenB.address, decimals: 18, vaultId: aliceOutputVault },
-      ],
-      evaluableConfig: EvaluableConfig_A,
-      meta: aliceOrder,
-    };
-
-    const txOrder_A = await orderBook.connect(alice).addOrder(orderConfig_A);
-
-    const {
-      order: Order_A
-    } = (await getEventArgs(
-      txOrder_A,
-      "AddOrder",
-      orderBook
-    ));
-
-     // DEPOSIT
-     // Deposit token equal to the size of the batch
-     const amountB = ethers.BigNumber.from("1000" + eighteenZeros); 
- 
-     const depositConfigStructAlice = {
-       token: tokenB.address,
-       vaultId: aliceOutputVault,
-       amount: amountB.mul(2),
-     };
- 
-     await tokenB.transfer(alice.address, amountB.mul(2));
-     await tokenB
-       .connect(alice)
-       .approve(orderBook.address, depositConfigStructAlice.amount);
- 
-     // Alice deposits tokenB into her output vault
-     await orderBook
-       .connect(alice)
-       .deposit(depositConfigStructAlice);
- 
-    // TAKE ORDER 0 
-     const takeOrderConfigStruct0 = {
-       order: Order_A,
-       inputIOIndex: 0,
-       outputIOIndex: 0,
-       signedContext : []
-     }; 
- 
-     let ratio0 = await prbScale(0,strategyRatio) 
-     
-     const takeOrdersConfigStruct0 = {
-       output: tokenA.address,
-       input: tokenB.address,
-       minimumInput: amountB,
-       maximumInput: amountB,
-       maximumIORatio: ratio0,
-       orders: [takeOrderConfigStruct0],
-     };
- 
-     const amountA0 = amountB.mul(ratio0).div(ONE); 
- 
-     await tokenA.transfer(bob.address, amountA0);
-     await tokenA.connect(bob).approve(orderBook.address, amountA0); 
-
-     const txTakeOrders0 = await orderBook
-        .connect(bob)
-        .takeOrders(takeOrdersConfigStruct0)
-
-      const { sender: sender0, config: config0, input:input0, output: output0 } = (await getEventArgs(
-        txTakeOrders0,
-        "TakeOrder",
-        orderBook
-      ));  
-
-
-      assert(sender0 === bob.address, "wrong sender");
-      assert(input0.eq(amountB), "wrong input");
-      assert(output0.eq(amountA0), "wrong output");
-
-      compareStructs(config0, takeOrderConfigStruct0); 
-
-     // TAKE ORDER 1  
-
-     const takeOrderConfigStruct1 = {
-      order: Order_A,
-      inputIOIndex: 0,
-      outputIOIndex: 0,
-      signedContext : []
-    }; 
-
-    let ratio1 = await prbScale(1,strategyRatio) 
-    
-    const takeOrdersConfigStruct1 = {
-      output: tokenA.address,
-      input: tokenB.address,
-      minimumInput: amountB,
-      maximumInput: amountB,
-      maximumIORatio: ratio1,
-      orders: [takeOrderConfigStruct1],
-    };
-
-    const amountA1 = amountB.mul(ratio1).div(ONE); 
-
-    await tokenA.transfer(bob.address, amountA1);
-    await tokenA.connect(bob).approve(orderBook.address, amountA1);  
-
-    await timewarp(43200);
-    
-    //Should fail as delay is not observed
-    await assertError(
-      async () =>
-        await orderBook
-      .connect(bob)
-      .takeOrders(takeOrdersConfigStruct0),
-      "",
-      "Delay"
-    )  
-
-    await timewarp(41400);
-    
-    //Should fail as delay is not observed
-    await assertError(
-      async () =>
-        await orderBook
-      .connect(bob)
-      .takeOrders(takeOrdersConfigStruct0),
-      "",
-      "Delay"
-    ) 
-    
-    // Introducing Delay
-    await timewarp(1800); 
-    
-    // Order should succeed after deplay is complete
-    const txTakeOrders1 = await orderBook
-        .connect(bob)
-        .takeOrders(takeOrdersConfigStruct1)
-
-      const { sender: sender1, config: config1, input:input1, output: output1 } = (await getEventArgs(
-        txTakeOrders1,
-        "TakeOrder",
-        orderBook
-      ));  
-
-
-      assert(sender1 === bob.address, "wrong sender");
-      assert(input1.eq(amountB), "wrong input");
-      assert(output1.eq(amountA1), "wrong output");
-
-      compareStructs(config1, takeOrderConfigStruct0); 
-        
-  }); 
-
-  it("should ensure ratio is not scaled by the expressions for orders with same batch", async function () { 
-
-    const signers = await ethers.getSigners();
-
-    const [, alice, bob] = signers;   
-
-    const aliceInputVault = ethers.BigNumber.from(randomUint256());
-    const aliceOutputVault = ethers.BigNumber.from(randomUint256());
-  
-
-    const aliceOrder = encodeMeta("Order_A"); 
-
-    // Order_A
-    const strategyRatio = "25e13"
-    const strategyExpression = path.resolve(
-      __dirname,
-      "../src/0-pilot.rain"
-    );
-
-    const strategyString = await fetchFile(strategyExpression); 
-
-    const stringExpression = mustache.render(strategyString, {
-      counterparty: bob.address,
-      ratio: strategyRatio
-    }); 
-
-    const { sources, constants } = await standardEvaluableConfig(stringExpression)
-
-    const EvaluableConfig_A = {
-      deployer: expressionDeployer.address,
-      sources,
-      constants,
-    }
-
-    const orderConfig_A= {
-      validInputs: [
-        { token: tokenA.address, decimals: 18, vaultId: aliceInputVault },
-      ],
-      validOutputs: [
-        { token: tokenB.address, decimals: 18, vaultId: aliceOutputVault },
-      ],
-      evaluableConfig: EvaluableConfig_A,
-      meta: aliceOrder,
-    };
-
-    const txOrder_A = await orderBook.connect(alice).addOrder(orderConfig_A);
-
-    const {
-      order: Order_A
-    } = (await getEventArgs(
-      txOrder_A,
-      "AddOrder",
-      orderBook
-    ));
-
-
-    // TAKE ORDER
-
-    // Bob takes order with direct wallet transfer
-    for(let i = 0 ; i < 10 ; i++){  
-
-      // DEPOSIT
-      const amountB = ethers.BigNumber.from("100" + eighteenZeros);
-
-      const depositConfigStructAlice = {
-        token: tokenB.address,
-        vaultId: aliceOutputVault,
-        amount: amountB,
-      };
-
-      await tokenB.transfer(alice.address, amountB);
-      await tokenB
-        .connect(alice)
-        .approve(orderBook.address, depositConfigStructAlice.amount);
-
-      // Alice deposits tokenB into her output vault
-       await orderBook.connect(alice).deposit(depositConfigStructAlice);
-
-      const takeOrderConfigStruct = {
-        order: Order_A,
-        inputIOIndex: 0,
-        outputIOIndex: 0,
-       signedContext : []
-      }; 
-
-      // Batch Index Remains the same hence ratio remains the same 
-      const ratio = await prbScale(0,strategyRatio)
-  
-      const takeOrdersConfigStruct = {
-        output: tokenA.address,
-        input: tokenB.address,
-        minimumInput: amountB,
-        maximumInput: amountB,
-        maximumIORatio: ratio,
-        orders: [takeOrderConfigStruct],
-      };
-  
-      const amountA = amountB.mul(ratio).div(ONE); 
-  
-      await tokenA.transfer(bob.address, amountA);
-      await tokenA.connect(bob).approve(orderBook.address, amountA); 
-  
-  
-      const txTakeOrders = await orderBook
-        .connect(bob)
-        .takeOrders(takeOrdersConfigStruct);   
-        
-      const { sender, config, input, output } = (await getEventArgs(
-        txTakeOrders,
-        "TakeOrder",
-        orderBook
-      ));  
-    
-      assert(sender === bob.address, "wrong sender");
-      assert(input.eq(amountB), "wrong input");
-      assert(output.eq(amountA), "wrong output");
-  
-      compareStructs(config, takeOrderConfigStruct); 
-
-      // No delay is introduced as all orders are part of the same batch
-      // await timewarp(86400)
-    } 
-    
-  });  
-
-  it("should ensure ratio is scaled exponentially by the expressions as batch index increases", async function () { 
-
-    const signers = await ethers.getSigners();
-
-    const [, alice, bob] = signers;   
-
-
-    const aliceInputVault = ethers.BigNumber.from(randomUint256());
-    const aliceOutputVault = ethers.BigNumber.from(randomUint256());
-  
-
-    const aliceOrder = encodeMeta("Order_A");  
-
-    const strategyRatio = "25e13"
-
-    // Order_A
-
-    const strategyExpression = path.resolve(
-      __dirname,
-      "../src/0-pilot.rain"
-    );
-
-    const strategyString = await fetchFile(strategyExpression); 
-
-    const stringExpression = mustache.render(strategyString, {
-      counterparty: bob.address,
-      ratio: strategyRatio
-    }); 
-
-    const { sources, constants } = await standardEvaluableConfig(stringExpression)
-
-    const EvaluableConfig_A = {
-      deployer: expressionDeployer.address,
-      sources,
-      constants,
-    }
-    const orderConfig_A= {
-      validInputs: [
-        { token: tokenA.address, decimals: 18, vaultId: aliceInputVault },
-      ],
-      validOutputs: [
-        { token: tokenB.address, decimals: 18, vaultId: aliceOutputVault },
-      ],
-      evaluableConfig: EvaluableConfig_A,
-      meta: aliceOrder,
-    };
-
-    const txOrder_A = await orderBook.connect(alice).addOrder(orderConfig_A);
-
-    const {
-      order: Order_A
-    } = (await getEventArgs(
-      txOrder_A,
-      "AddOrder",
-      orderBook
-    ));
-
-
-    // TAKE ORDER //255000000000000000
-
-    // Bob takes order with direct wallet transfer
-    for(let i = 0 ; i < 10 ; i++){   
-
-      // DEPOSIT
-
-      // Deposit max amount per batch
-      const amountB = ethers.BigNumber.from("1000" + eighteenZeros);
-
-      const depositConfigStructAlice = {
-        token: tokenB.address,
-        vaultId: aliceOutputVault,
-        amount: amountB,
-      };
-
-      await tokenB.transfer(alice.address, amountB);
-      await tokenB
-        .connect(alice)
-        .approve(orderBook.address, depositConfigStructAlice.amount);
-
-      // Alice deposits tokenB into her output vault
-       await orderBook.connect(alice).deposit(depositConfigStructAlice);
-
-      const takeOrderConfigStruct = {
-        order: Order_A,
-        inputIOIndex: 0,
-        outputIOIndex: 0,
-       signedContext : []
-      }; 
-
-      // Scaling ratio as batch index increases 
-      const ratio = await prbScale(i,strategyRatio) 
-
-      const takeOrdersConfigStruct = {
-        output: tokenA.address,
-        input: tokenB.address,
-        minimumInput: amountB,
-        maximumInput: amountB,
-        maximumIORatio: ratio,
-        orders: [takeOrderConfigStruct],
-      };
-     
-  
-      const amountA = amountB.mul(ratio).div(ONE)
-  
-      await tokenA.transfer(bob.address, amountA);
-      await tokenA.connect(bob).approve(orderBook.address, amountA); 
-  
-  
-      const txTakeOrders = await orderBook
-        .connect(bob)
-        .takeOrders(takeOrdersConfigStruct);   
-        
-      const { sender, config, input, output } = (await getEventArgs(
-        txTakeOrders,
-        "TakeOrder",
-        orderBook
-      ));   
-
-    
-      assert(sender === bob.address, "wrong sender");
-      assert(input.eq(amountB), "wrong input");
-      assert(output.eq(amountA), "wrong output");
-  
-      compareStructs(config, takeOrderConfigStruct); 
-
-      // Delay is introduced between batches
-      await timewarp(86400)
-    } 
-    
-  });  
-
-  it("should ensure batch info is not shared across orders in case of multiple orders by a same user ", async function () { 
-
-    const signers = await ethers.getSigners();
-
-    const [, alice, bob] = signers;   
-
-
-    const aliceVaultA = ethers.BigNumber.from(randomUint256());
-    const aliceVaultB = ethers.BigNumber.from(randomUint256());
-    const aliceVaultC = ethers.BigNumber.from(randomUint256());
-
-  
-
-    const aliceOrder = encodeMeta("Order_A");  
-    
-    const strategyExpression = path.resolve(
-      __dirname,
-      "../src/0-pilot.rain"
-    );
-
-    const strategyString = await fetchFile(strategyExpression); 
-
-   
-    // Order_A 
-
-    const strategyRatio_A = "25e13"
-
-    
-    const stringExpression_A = mustache.render(strategyString, {
-      counterparty: bob.address,
-      ratio: strategyRatio_A
-    }); 
-
-    const { sources, constants } = await standardEvaluableConfig(stringExpression_A)
-
-    const EvaluableConfig_A = {
-      deployer: expressionDeployer.address,
-      sources,
-      constants,
-    }
-    const orderConfig_A= {
-      validInputs: [
-        { token: tokenA.address, decimals: 18, vaultId: aliceVaultA },
-      ],
-      validOutputs: [
-        { token: tokenB.address, decimals: 18, vaultId: aliceVaultA },
-      ],
-      evaluableConfig: EvaluableConfig_A,
-      meta: aliceOrder,
-    };
-
-    const txOrder_A = await orderBook.connect(alice).addOrder(orderConfig_A);
-
-    const {
-      order: Order_A
-    } = (await getEventArgs(
-      txOrder_A,
-      "AddOrder",
-      orderBook
-    )); 
-
-    // Order B 
-    // Placing order with diff ratio
-    const strategyRatio_B = "30e13" 
-
-    const stringExpression_B = mustache.render(strategyString, {
-      counterparty: bob.address,
-      ratio: strategyRatio_B
-    }); 
-
-    const { sources:sourceB, constants:constantsB } = await standardEvaluableConfig(stringExpression_B)
-
-    const EvaluableConfig_B = {
-      deployer: expressionDeployer.address,
-      sources: sourceB,
-      constants: constantsB,
-    }
-    const orderConfig_B= {
-      validInputs: [
-        { token: tokenA.address, decimals: 18, vaultId: aliceVaultB },
-      ],
-      validOutputs: [
-        { token: tokenB.address, decimals: 18, vaultId: aliceVaultB },
-      ],
-      evaluableConfig: EvaluableConfig_B,
-      meta: aliceOrder,
-    };
-
-    const txOrder_B = await orderBook.connect(alice).addOrder(orderConfig_B);
-
-    const {
-      order: Order_B
-    } = (await getEventArgs(
-      txOrder_B,
-      "AddOrder",
-      orderBook
-    ));  
-
-    // Order C
-    // Placing order with exact params as order A.
-    // Even if the params are same order hash computed is different
-    const strategyRatio_C = "25e13" 
-
-    const stringExpression_C = mustache.render(strategyString, {
-      counterparty: bob.address,
-      ratio: strategyRatio_C
-    }); 
-
-    const { sources:sourceC, constants:constantsC } = await standardEvaluableConfig(stringExpression_C)
-
-    const EvaluableConfig_C = {
-      deployer: expressionDeployer.address,
-      sources: sourceC,
-      constants: constantsC,
-    }
-    const orderConfig_C= {
-      validInputs: [
-        { token: tokenA.address, decimals: 18, vaultId: aliceVaultC },
-      ],
-      validOutputs: [
-        { token: tokenB.address, decimals: 18, vaultId: aliceVaultC },
-      ],
-      evaluableConfig: EvaluableConfig_C,
-      meta: aliceOrder,
-    };
-
-    const txOrder_C = await orderBook.connect(alice).addOrder(orderConfig_C);
-
-    const {
-      order: Order_C
-    } = (await getEventArgs(
-      txOrder_C,
-      "AddOrder",
-      orderBook
-    ));
- 
-
-    // TAKE ORDER 
-
-    // Bob takes order with direct wallet transfer for order A and order B
-    for(let i = 0 ; i < 10 ; i++){   
-
-        // Since orders are different orders , bob should be able to take orders concurrently/with delay 
-        await takeOrder(
-          alice, 
-          bob, 
-          tokenA,
-          tokenB, 
-          aliceVaultA,
-          Order_A,
-          orderBook,
-          i,
-          strategyRatio_A
-        )  
-        // placing delay
-        await timewarp(3600) 
-
-        await takeOrder(
-          alice, 
-          bob, 
-          tokenA,
-          tokenB, 
-          aliceVaultB,
-          Order_B,
-          orderBook,
-          i,
-          strategyRatio_B
-        )   
-
-        // placing delay
-        await timewarp(3600) 
-
-        await takeOrder(
-          alice, 
-          bob, 
-          tokenA,
-          tokenB, 
-          aliceVaultC,
-          Order_C,
-          orderBook,
-          i,
-          strategyRatio_C
-        )
-
-      // Delay is introduced between batches
-      await timewarp(86400)
-    }   
-
-       
-  });  
-
-  it("should ensure there is no overflow is allowed at end of batch", async function () { 
-
-    const signers = await ethers.getSigners();
-
-    const [, alice, bob] = signers;   
-
-
-    const aliceInputVault = ethers.BigNumber.from(randomUint256());
-    const aliceOutputVault = ethers.BigNumber.from(randomUint256());
-  
-
-    const aliceOrder = encodeMeta("Order_A"); 
-
-    // Order_A
-    const strategyRatio = "25e13"
-    const strategyExpression = path.resolve(
-      __dirname,
-      "../src/0-pilot.rain"
-    );
-
-    const strategyString = await fetchFile(strategyExpression); 
-
-    const stringExpression = mustache.render(strategyString, {
-      counterparty: bob.address,
-      ratio: strategyRatio
-    }); 
-
-    const { sources, constants } = await standardEvaluableConfig(stringExpression)
-
-    const EvaluableConfig_A = {
-      deployer: expressionDeployer.address,
-      sources,
-      constants,
-    }
-    const orderConfig_A= {
-      validInputs: [
-        { token: tokenA.address, decimals: 18, vaultId: aliceInputVault },
-      ],
-      validOutputs: [
-        { token: tokenB.address, decimals: 18, vaultId: aliceOutputVault },
-      ],
-      evaluableConfig: EvaluableConfig_A,
-      meta: aliceOrder,
-    };
-
-    const txOrder_A = await orderBook.connect(alice).addOrder(orderConfig_A);
-
-    const {
-      order: Order_A
-    } = (await getEventArgs(
-      txOrder_A,
-      "AddOrder",
-      orderBook
-    ));
-
-
-    // TAKE ORDER
-
-    // Bob takes order with direct wallet transfer
-    for(let i = 0 ; i < 3 ; i++){  
-
-      // DEPOSIT
-
-      // Deposit max amount per batch
-      
-      const amountB = i == 0 ? (
-        ethers.BigNumber.from("1000" + eighteenZeros)
-      ) : (
-        i == 1 ? (
-          ethers.BigNumber.from("999" + eighteenZeros)
-        ) : ( 
-          // Since overflow is set to 0 we cannot deposit more than
-          // remaining amount for the batch
-          ethers.BigNumber.from("1" + eighteenZeros)
-        )
-      );
-
-      const depositConfigStructAlice = {
-        token: tokenB.address,
-        vaultId: aliceOutputVault,
-        amount: amountB,
-      };
-
-      await tokenB.transfer(alice.address, amountB);
-      await tokenB
-        .connect(alice)
-        .approve(orderBook.address, depositConfigStructAlice.amount);
-
-      // Alice deposits tokenB into her output vault
-       await orderBook.connect(alice).deposit(depositConfigStructAlice);
-
-      const takeOrderConfigStruct = {
-        order: Order_A,
-        inputIOIndex: 0,
-        outputIOIndex: 0,
-       signedContext : []
-
-      }; 
-
-      // Scaling ratio as batch index increases 
-      const ratio =  i == 0 ? (
-        await prbScale(0,strategyRatio)
-      ) : (
-        await prbScale(1,strategyRatio)
-      )
-  
-      const takeOrdersConfigStruct = {
-        output: tokenA.address,
-        input: tokenB.address,
-        minimumInput: amountB,
-        maximumInput: amountB,
-        maximumIORatio: ratio,
-        orders: [takeOrderConfigStruct],
-      };
-  
-      const amountA = amountB.mul(ratio).div(ONE); 
-  
-      await tokenA.transfer(bob.address, amountA);
-      await tokenA.connect(bob).approve(orderBook.address, amountA); 
-  
-  
-      const txTakeOrders = await orderBook
-        .connect(bob)
-        .takeOrders(takeOrdersConfigStruct);   
-        
-      const { sender, config, input, output } = (await getEventArgs(
-        txTakeOrders,
-        "TakeOrder",
-        orderBook
-      ));   
-
-    
-      assert(sender === bob.address, "wrong sender");
-      assert(input.eq(amountB), "wrong input");
-      assert(output.eq(amountA), "wrong output");
-  
-      compareStructs(config, takeOrderConfigStruct); 
-
-      // Delay is introduced between batches
-      await timewarp(86400)
-    }  
-
-    // Bob takes order with direct wallet transfer
-    for(let i = 0 ; i < 3 ; i++){  
-
-      // DEPOSIT
-
-      // Deposit max amount per batch
-      
-      const amountB = i == 0 ? (
-        ethers.BigNumber.from("1000" + eighteenZeros)
-      ) : (
-        i == 1 ? (
-          ethers.BigNumber.from("999" + eighteenZeros)
-        ) : ( 
-          // Setting deposit amount more than remaining amount
-          // for batch
-          ethers.BigNumber.from("10" + eighteenZeros)
-        )
-      );
-
-      const depositConfigStructAlice = {
-        token: tokenB.address,
-        vaultId: aliceOutputVault,
-        amount: amountB,
-      };
-
-      await tokenB.transfer(alice.address, amountB);
-      await tokenB
-        .connect(alice)
-        .approve(orderBook.address, depositConfigStructAlice.amount);
-
-      // Alice deposits tokenB into her output vault
-       await orderBook.connect(alice).deposit(depositConfigStructAlice);
-
-      const takeOrderConfigStruct = {
-        order: Order_A,
-        inputIOIndex: 0,
-        outputIOIndex: 0,
-       signedContext : []
-
-      }; 
-
-      // Scaling ratio as batch index increases 
-      const ratio =  i == 0 ? (
-        await prbScale(i+2,strategyRatio)
-      ) : ( 
-        // check for the third batch
-        await prbScale(3,strategyRatio)
-      )
-  
-      const takeOrdersConfigStruct = {
-        output: tokenA.address,
-        input: tokenB.address,
-        minimumInput: amountB,
-        maximumInput: amountB,
-        maximumIORatio: ratio,
-        orders: [takeOrderConfigStruct],
-      };
-  
-      const amountA = amountB.mul(ratio).div(ONE); 
-  
-      await tokenA.transfer(bob.address, amountA);
-      await tokenA.connect(bob).approve(orderBook.address, amountA); 
-  
-      if(i == 0 || i == 1  ){ 
-
-        const txTakeOrders = await orderBook
-        .connect(bob)
-        .takeOrders(takeOrdersConfigStruct);   
-        
-        const { sender, config, input, output } = (await getEventArgs(
-          txTakeOrders,
-          "TakeOrder",
-          orderBook
-        ));   
-
-      
-        assert(sender === bob.address, "wrong sender");
-        assert(input.eq(amountB), "wrong input");
-        assert(output.eq(amountA), "wrong output");
-    
-        compareStructs(config, takeOrderConfigStruct);
-
-      }else{ 
-        await assertError(
-          async () =>
-             await orderBook
-            .connect(bob)
-            .takeOrders(takeOrdersConfigStruct),
-          "",
-          "Overflow"
-        );
-      }
-       
-      // Delay is introduced between batches
-      await timewarp(86400)
-
-    } 
-    
-  }); 
-
-
-
   describe("should chain orders within same batch with decimals", () => { 
 
     it("should ensure ratio is not scaled based on input/output token decimals: (Input Decimals: 6 vs Output Decimals: 18)", async function () { 
@@ -1097,7 +106,7 @@ describe("Pilot", async function () {
       const strategyRatio = "25e13"
     const strategyExpression = path.resolve(
       __dirname,
-      "../src/0-pilot.rain"
+      "../src/1-in-token-batch.rain"
     );
 
     const strategyString = await fetchFile(strategyExpression); 
@@ -1234,7 +243,7 @@ describe("Pilot", async function () {
       const strategyRatio = "25e13"
     const strategyExpression = path.resolve(
       __dirname,
-      "../src/0-pilot.rain"
+      "../src/1-in-token-batch.rain"
     );
 
     const strategyString = await fetchFile(strategyExpression); 
@@ -1370,7 +379,7 @@ describe("Pilot", async function () {
       const strategyRatio = "25e13"
       const strategyExpression = path.resolve(
         __dirname,
-        "../src/0-pilot.rain"
+        "../src/1-in-token-batch.rain"
       );
 
       const strategyString = await fetchFile(strategyExpression); 
@@ -1507,7 +516,7 @@ describe("Pilot", async function () {
       const strategyRatio = "25e13"
     const strategyExpression = path.resolve(
       __dirname,
-      "../src/0-pilot.rain"
+      "../src/1-in-token-batch.rain"
     );
 
     const strategyString = await fetchFile(strategyExpression); 
@@ -1615,10 +624,10 @@ describe("Pilot", async function () {
 
   }) 
 
-  describe("should scale ratio exponentially for different batches with decimals", () => { 
+  describe.only("should scale ratio exponentially for different batches with decimals", () => { 
 
     // Scaling ratio based on FixedPointMath `scaleRatio`
-    xit("should ensure ratio is scaled exponentially based on input/output token decimals: (Input Decimals: 6 vs Output Decimals: 18)", async function () { 
+    it.only("should ensure ratio is scaled exponentially based on input/output token decimals: (Input Decimals: 6 vs Output Decimals: 18)", async function () { 
 
       const tokenA06 = (await basicDeploy("ReserveTokenDecimals", {}, [
         6,
@@ -1649,7 +658,7 @@ describe("Pilot", async function () {
       const strategyRatio = "25e13"
       const strategyExpression = path.resolve(
         __dirname,
-        "../src/0-pilot.rain"
+        "../src/1-in-token-batch.rain"
       );
 
       const strategyString = await fetchFile(strategyExpression); 
@@ -1692,10 +701,15 @@ describe("Pilot", async function () {
       // Recursively places orders for batches
       for(let i = 0 ; i < 10 ; i++){  
   
-        // DEPOSIT
+        // DEPOSIT 
+        console.log(`------------------------- i : ${i}`)
+        // scale ratio as batch index increases
+        let ratio = await prbScale(i,strategyRatio)  
+        console.log("ratio : " , ratio)
 
-        // Deposit amount same as max posit amount per batch
-        const amountB = ethers.BigNumber.from("1000" + eighteenZeros);
+        // Deposit max amount per batch 
+        const amountB = await scaleOutputMax(ratio.toString(),18)  
+        console.log("amountB : " , amountB )
   
         const depositConfigStructAlice = {
           token: tokenB18.address,
@@ -1706,7 +720,7 @@ describe("Pilot", async function () {
         await tokenB18.transfer(alice.address, amountB);
         await tokenB18
           .connect(alice)
-          .approve(orderBook.address, depositConfigStructAlice.amount);
+          .approve(orderBook.address, max_uint256);
   
         // Alice deposits tokenB into her output vault
          await orderBook.connect(alice).deposit(depositConfigStructAlice);
@@ -1718,9 +732,6 @@ describe("Pilot", async function () {
           signedContext : []
         }; 
         
-        // scale ratio as batch index increases
-        let ratio = await prbScale(i,strategyRatio) 
-      
         const maximumIORatio = await scaleRatio(ratio,tokenADecimals,tokenBDecimals)
     
         const takeOrdersConfigStruct = {
@@ -1728,11 +739,12 @@ describe("Pilot", async function () {
           input: tokenB18.address,
           minimumInput: amountB,
           maximumInput: amountB,
-          maximumIORatio: maximumIORatio,
+          maximumIORatio: max_uint256,
           orders: [takeOrderConfigStruct],
         };
     
-        const amountA = amountB.mul(maximumIORatio).div(ONE) 
+        const amountA = ethers.BigNumber.from('1005000000')
+        console.log("amountA : " , amountA )
         
         await tokenA06.transfer(bob.address, amountA);
         await tokenA06.connect(bob).approve(orderBook.address, amountA); 
@@ -1746,21 +758,17 @@ describe("Pilot", async function () {
           txTakeOrders,
           "TakeOrder",
           orderBook
-        ));  
-    
-        assert(sender === bob.address, "wrong sender");
-        assert(input.eq(amountB), "wrong input");
-        assert(output.eq(amountA), "wrong output");
-    
-        compareStructs(config, takeOrderConfigStruct); 
+        )); 
+        
+        console.log("input : " , input )
+        console.log("output : " , output )
 
-        //Introduce delay
         await timewarp(86400);
       } 
       
     });  
 
-    it("should ensure ratio is scaled exponentially based on input/output token decimals: (Input Decimals: 18 vs Output Decimals: 6)", async function () { 
+    xit("should ensure ratio is scaled exponentially based on input/output token decimals: (Input Decimals: 18 vs Output Decimals: 6)", async function () { 
 
       const tokenA18 = (await basicDeploy("ReserveTokenDecimals", {}, [
         18,
@@ -1790,7 +798,7 @@ describe("Pilot", async function () {
       const strategyRatio = "25e13"
       const strategyExpression = path.resolve(
         __dirname,
-        "../src/0-pilot.rain"
+        "../src/1-in-token-batch.rain"
       );
 
       const strategyString = await fetchFile(strategyExpression); 
@@ -1831,10 +839,17 @@ describe("Pilot", async function () {
   
   
       // Recursively places orders for batches
-      for(let i = 0 ; i < 10 ; i++){  
+      for(let i = 0 ; i < 10 ; i++){   
+        
+        console.log("---------------- i : " , i)
+
+        // scale ratio as batch index increases
+        let ratio = await prbScale(i,strategyRatio) 
   
         // DEPOSIT
-        const amountB = ethers.BigNumber.from("1000" + sixZeros);
+        // Deposit max amount per batch 
+        const amountB = await scaleOutputMax(ratio.toString(),6)  
+        console.log("amountB : " , amountB )
   
         const depositConfigStructAlice = {
           token: tokenB06.address,
@@ -1857,9 +872,7 @@ describe("Pilot", async function () {
           signedContext : []
         }; 
         
-        // scale ratio as batch index increases
-        let ratio = await prbScale(i,strategyRatio) 
-      
+        
         const maximumIORatio = await scaleRatio(ratio,tokenADecimals,tokenBDecimals)
     
         const takeOrdersConfigStruct = {
@@ -1871,7 +884,8 @@ describe("Pilot", async function () {
           orders: [takeOrderConfigStruct],
         };
     
-        const amountA = amountB.mul(maximumIORatio).div(ONE) 
+        const amountA = ethers.BigNumber.from('1000' + eighteenZeros) 
+      
         
         await tokenA18.transfer(bob.address, amountA);
         await tokenA18.connect(bob).approve(orderBook.address, amountA); 
@@ -1887,11 +901,11 @@ describe("Pilot", async function () {
           orderBook
         ));  
     
-        assert(sender === bob.address, "wrong sender");
-        assert(input.eq(amountB), "wrong input");
-        assert(output.eq(amountA), "wrong output");
+        // assert(sender === bob.address, "wrong sender");
+        // assert(input.eq(amountB), "wrong input");
+        // assert(output.eq(amountA), "wrong output");
     
-        compareStructs(config, takeOrderConfigStruct); 
+        // compareStructs(config, takeOrderConfigStruct); 
 
         //Introduce delay
         await timewarp(86400);
@@ -1930,7 +944,7 @@ describe("Pilot", async function () {
       const strategyRatio = "25e13"
       const strategyExpression = path.resolve(
         __dirname,
-        "../src/0-pilot.rain"
+        "../src/1-in-token-batch.rain"
       );
 
       const strategyString = await fetchFile(strategyExpression);
