+++ utils/interpreter/interpreter.ts
@@ -17,7 +17,7 @@ import { MAGIC_NUMBERS, decodeRainMetaDocument } from "../meta/cbor";
 export const standardEvaluableConfig = async (
   expression: string
 ): Promise<ExpressionConfig> => {
-  const rainDocumentEncoded = "0xff0a89c674ee7874a40059111a789ced5d5b93dbb6157ecfafc0ec78a6762ada2b6dec6cfd96d869eb696c671ca79d8ee307888424cc52844c90da553af9ef3de700bc1312745befeee8c15e490409e03bdfb90038003f7dc3d8ffe01f6367099f8bb397ec2c12a18a44307afee26c60ae44428778e5355d619ce9588682a9091bcb4c339567f89933b8057f614b1ee7823de6c92a9bc964ca54c2b299603ae3e1d593a7c553d542a43c89e0c19fe887a2218dc6c03d6966ef68b4e547a8472691b8619962548aa5824758dd2455f3a7f57bb09558cde5800d9f7fb6bfff3970d71a8b649acdfaaa7d97cfc7222dbb0e7563ad54a569456fc5e703f67d592ffdfd6c4190c922a732452bce163c85566422d53564eaadaca3a3f2341467e5a53fed2753d79f05d079662b19da5f782cb91654c119c992a48d77d13d2d4288c445889fe80ad36a2e5896e649c83311196408114f4a80200148ce1271ddbce1c495237265b0be3c34702ab2fdb815aac52af020d8cd2246865d8cda04fb0565c5139603439019864644974ba605b222136cc8c62bf8a3e51fc294d03dc4393f088af4787f502e7b40b1bdc5ce965d774213ceb84c62995c058b14606ce3f3915f09cdfe2e40eba097ecd78cc7e2870974c5e250922dbca20220514d8af7aa782e7b9ff23016f48cbf6846b550517313964d85cee38c8da107112a6e55cbb1509e40636aba51eb308f22688e46a5c2a699e652716f626b6c7ec06bed6fd511e5c02a091dc54a249ab619cf6c55d7328ed9986826a3fd94c308d429789564e2266b0bfc4da2450a32049db0255828a049a412a545057d282d5ca8e27c9e904853756d0851c86b1beb6a9eb397a984fa77b478e75d44d7e31698e60633ae676d0c7f45888c1ae0658a5d1226924ca6a24295ee7fca7ed36292c76ca252a613bed0339565082be21973d4b47026c22bfa89c55253289483904061d55246a0324e93b413dae70db40f0557259a2ed57c99863f5ab6d559d6a41eddb52714450b9bb2ba4d678a70f92bbf03fb898aa3c0a1e7bf8814283747f4630530452ce400be5a02e128dc32ed37961abf175840cb34d30b11ca89c4586cc5b01a43cfc214681b04d9ef2634c2275dcb289bc1f74a74e9349f836668bcc9443ac86f9909681a94c3dab612a685bd4754c391af19a156f6f12129032ed3338ab9103ec2cd65ba863e755a59b99948484275e82549134cbdbdacfcaececa35959a902d20f9f4d56a04675da1294c5166eed2850b5f5d6809a578da563a424f6d7a58bd405b011761b020bcf5c7d5aa46f1ba674561b91d2be8517f1865f48aa84da6aa327264c1f204b5d1c27c2d4113743e0eaa02a48f55bca4810a1840d58ce576a399830adf9be806f60063e9f52a56c8670f7ebb8dc12e5ca57bee18532b247a790a23660c34712030499c8c8dc4389fb629fbc38f6f9819981bd2d9f0a5a263aca660a2b38280339e4618c6829fd02a165b51718e23941e3250c3a6e85bda256a3239df3a5639ef810a7a832895ad2f3f0678c5899c0aae6732ee0c9b3e185da66b85a16615b58d0dc8c02824c53c86c135079063131045122507086b0634bb0f9abd41db6abfd3e0e6034fa678df271b71d60b846a0ecfa241120dcad2290bd8f06cbda037286ceb993465f25778e891357ac360b19474ff50b1bc8ce2aae8e41f1bf683e2b41886cdee791440276993dd6390dea3838e2ab07741a78a9f51854adf98d0a85963ec093a84e6a1ed591b4e332bc60e0dbf49c38cedbd26b56fb3f3c2627b46a127fffce0fd33a932fa66953a750e9b11ccc55ca5abfea89227c47c437c745a19a7f1549af21553695d218c86a05a481a7b339ab7de3982f4a32635cece4df76a84af4264ab452f25f1f7a232ac64d08602a7566a9d6b9366fb00a26fb68dbae71262df2cd13f716ae8318418122cfc8a5d8930e457a3e72f9ed849e85a785099b3f6a4e981c6c6348de176d020a2d173f0d0db2be4f75f59215d5323220d83e1f0f9f360cc639e80995593b67cfe213202da96b0b3783f7d7885f701a1513ae6371e860a22b72d15eb3033d9d48ec0ce5afbcf51db266fb15a43f5ec3b25dd86dd1d6974e5031fb3b0a3452d29e95e31e1e42add6daf9bce1b4b684ac8e8a076714f3d7347c28f6d28fc843d63a3fe78787decd7f246becadb6e0645cfdfb2d193dd43e83d997b80d8bbe4f4ce86c6c16d609a17c347e7bbd89fd1f9c3b33e3bda920ac04d38672ae371a0f3c522eec45105d2e66a0fd07709dadd91b29ddf0054b10855b7bcbcb37c5042664b6fcb52c4e0eb63796827596277083fd9230b6fd9d5d9ee23bd7bc1fcaf05be87de3cbf185e04ea3aa932161a6bc184195dae20c65b18ba55a839a169d32ce56176b4c492a2821d4cca9a8e7f3f1aeee2c2e0b6930fb354ab41b89663588e38b406e816c71e00cc871b7794f8adc943d479da9dcb175f725cf9a0c993c7fa4939e596d9e5b731e5268a01934fc5533685e811b336b21900fe8748d50086ea4b827ac212751ab16f105fdfe24c6a44e014dc3856e15560a6265d5926943c92a7a900fe53f9623d666bfafbcd07d9ba02aa6b43cb693a3be3f3457fe37bda5edee1ddfc6e6b3b1daa373fc17c1d56f602eb6b7f372d76f50c347d9d286cce1fe892cee7f8e75d9151589bf7aad4ec6b8cc647f756877ad3b0f3b97b19532efd64050565244ee23abeb80cd2eb329afd24060555826bf76609f124b6e38a6d61cc26fc591367ccf9cd3ae1c165393f19c5db1097851a45061f31b076ca4c7616a51b3293c94966b724330335c94c266b65a6d6862070f924afdb596e9be79d94c88624c059c9452c437ef253b727149d8fd709052ed384d14922b765d672ed34653ac43d45c3cbeee4045d619c4de48d88d84281246a9986ef58244239e731a5aa0f2fabafd733910ab82e29cf9e268f8cb4b6db47523caf2f31e37551979d3131790d6381f958a96978d4bb45b19914de953c260cbeb8b8bca8643f70b750f30cf75bf5a68ebc99b021c3866093f285d97d55dcc0e0bf5931938309fe93585db305d759b95d108286de0e604ecb86f67bb53d55791205f96263db23759d98d6d32dd8977aebe932e831b035941a353a565abbb6879e7b377d83ca1d71e363ef101719053ab249898268053fc9f028cad4671eb7dc667c22ec7d206c3115565ac08abb83dba27ac56476365994dfd62b40672053f23e4152bf7fff56f72900a6cb52b177dbd1d9ecf30ed6398a8f334c77c3620d0d33d565eae4234e2a57539c460056a790cd9bac88c3deedbfcac9934ebe483d4a9ec814649da824c0b51edb045c6d0398cb9887521d681bafc655a08cbc85b8911a935313316022d6c2aec8695a346a05d4db69dc29a2261bb926a95af0a94803d959473519d517ee4d7a24442b432b7bca6f06c75f7280929eeb032981bb386cb986585b82c7df2f5253f2886900ebf69b7ca42dd6c59e1368ab4002a365f4f66e60fc34ee5a844e88769e5ebfb69a11e4c4eeef2e2aaf41fa0cb9e0df820997f1815a40b54fc01fede9afa57b964a7cc9791c64aacdc5576a0e82b5bbef62f42823cbcbd60234b84d34e2d5a2be26dd7936aced71b19d5b318e9627ffc2e316f7a88fcfce8f453b827ab8658834da73d1ff8b1b73f0d41e66ddcb9ad3b67942774ee816a43d59f5daa5c35975c4c925569b7b11603cd36fdaad1ef51f5e633658d754a736bf35e9d002ac4323d7a366ea4d017ca4552b9bfb71dc53836e59bda699530e529306742c5a3d02aa5024cd7059af86083059497f05c3b5a7edd781f1203048c30420176ab100c7795bd4c5ca0ec1db3569257788c98d4d776e092cd27190472fce6f6064cd979d2df1bf7cf8f12d0c98d86fafa9c84bf68ac7618e87f418a7c15399cda0a332641c1c0cc498e818ccc927ab8119751503b1a399821b7fec56fb9e2f353628d1dc037ef2c13514b2b316d401f6bf52c4918d9ee7e0600550f77aa66050678b1493120d0bac5246c114867437b702f08ea8190c0836fae8835b4f8e4e07b6d7943402b85dab16501ab310b33c4dcc413d781062b3c053f601e989f32ed73c8d8c557e202445e87c20ee49aad9a4f209cd03c5659a0d29fc5da61ef59198879f3c51e91e9bb80196b14c78ba6aa2c2725d6cc4b79727c5f222f47fa6fa169eee126ca30ab79117709358a94e3aea3a53678c99cbd4551efbded8398b00c1663e7be106b4d806367113223463b1c21886621dacebceaba2e9a701073ffa60339d6fab8a53a1a04f29c42473c193464042f9ea8f1ee92f69f6f8867dcb564f1e3d7a9861ca746e819eba9360eb30cb64b3b3ade13c64cfd84d0d39e3436fcf85ee880af59260c14f3eb8c49db189af838cd5d484c7775e2de3c48212bb4fcc6a60a2a6c3f36d6109d57c4e2b2cf70615d34b030c7df6c466d7d8e15e4133aa90f18b0d7a12e23ab8bcb55971fd017d39f1bb31b27f18261c11f34176d13d68b683ec072ee1d1eca638bb8ed2b59167ab07021682e00956ee8fd663881f9adc7ad24470c51ee78996d3040f86e51a620e9964628ae5aae07fc2e72ad78cc75365f4fbf7b3729382c9ff1caf9886f836853b7e3f7b3812c9bd4482d1d8b62693d012107f2833d4fa1a41dc8eb898ee92f1a48fced4159370015d0ad6ec2cb37bc90893f28862e762ae9c60b2c352a478168cd9a459ce4edb1d345385276c8dd592b665942f8fa89d8e765abcd942e2204523be8640b5c9e6c78beb76a9d5ee58934f5ee49283e0ab5d502742dc6142acf3ec0d9638d3d55ba9ea0796f61853baeaeb5927016f2760949c73a950a59148c74a5d05cb61b0e4807a791a424bda1f3b274994f732baf1586eae6a62717c4103ac7afb5aaf0f79f3beb8f5df23ff70c39cf2b06d5def8bb31fcce22ae2e15d63f37006df1a3fe25ddb2cd5a170f73dcaa14591ba749a97dc8bd39804bb1c05a9987389c3a6c01e32912c2106ed3b67930c822dccd28fff7a87eba71373f4359dffa0f140f823b1cfb4f7200741d193dcddf6400cecedb25731eda5e2d8a00693ee0124a65f1b01a0bf985996eb7e5f54b81828605e59f72b66a21667eff4bcdbc41e608cc7e915dad57455f705438bca46088920bd0c02bdba6fc4319df1e8331ec76559067f432197a23388a12c782cc970846c56cd1a6a85518ab99598830ce17d27eddf55ac7a41708207f1989cac302ab0c716598a05bc388ea409df077cf79dcd8569ab6175f451a95ce6f16c397469e7ede5bcd8ae624f773c85ca574c0d476af0dda0b753d1998d3040b32bb17a661286165ca6e599cb064a0d0e05f35f80a7e63555c5eccf15e6a0360e32ae27f51f9cc4505d7f3853b4c58632cdceecbf6b407761fb4f0a23912e6ebdc73c5f5359e2631bd26322b54528b755565cdf0143d445b7f6671242600a3916aa3b21d63ceb044bb4d49c7c2f3ea3c20f71550b1c23f2385e15efb3c2bd7b329c51e6eeddd9d9fc79c03e5de07fc3f3cdefc5709c03bce3d1bef6bd18a3b2d9db993292dbf14fda6cbed5cc5e6c6d62c00298256132268f35c0a50e5b967ab299dc5730516980bfafdd835fce62109763b114f143243c71fd3b22fc7033e12f8e41f88b3d080f92bd8513d8912bf75b415ac4f798f48be4a4b349aba12065518645452aececd0c8aa89d77915266cc42940f754e058b8a701ef6f123f4ecf11c64e598858845910679d78db719a1f952fa2bd2a67ad4c58ab4c9aeb351773057757bbee0b39422889efaa9c627635880bdf08b68fada227f5ed99353d60d06353db4b46bb8530eb8e272bd6b9afdccf7b3160c3eaa5350377d5aed79dd56ac6222ff154a9012d33e06b5c307fbfafdecb01fbdbfacdb823af7dc4f7e7acd15480a293a6774ece6c214a43eaf2a84c8ca669ff66463f03bd0a6ef129c743c9fd75af1196f6546a7d34f29d84fa33c814397e1023eb50d57c11f1cc42521ad920253979e92e3da07c5916de67de508dd926145f14bd6aab2ca966479d0f12528824aa47498d5e50f2a7cdb893e6840283411f89bf5b4f6260e3a5df5101f866dbb54da2127e8ddafc62bccba36ca4af53c5ad3d595b7bcc6d5675881f993a8cce640d9db1f5606056d5b59fa32b7b8cf80c58fdbb5511f280df7cfee6ffb520bde3011bffe5282f43e495b402706170706c69636174696f6e2f6a736f6e03676465666c617465";
+  const rainDocumentEncoded = "0xff0a89c674ee7874a400591149789ced5deb8fdb3612ffdebf825804b86ccfdaacbd4dbbd76f7ddc23b83629d25c0f87b407d0126d132b898e2879d73df47fbf9921f51665f9b52ff843b2bb12250e7ff39be1901c521f3f63ec7ff08fb1b39847e2ec6b7616085f05c29bbcfef26c64ee0442fb78e77bbac338d3a1f40553333695a9662a4bf177cee011bcc2563ccc047bc9e375ba90f19ca998a50bc174cafd9bf38bfcad6a29121e07f0e28f742117a4260c3c93a4f6899a2cdf423d320ec41d4b15a3522c113cc0ea66898a2eaacfa09458cdf5888d5fff66afff3172d71a8a789e2ebaaa7d9b455391144d87bab156aad248d159f1e5887d55d44b3f7fb320c8789951995c8ab3254f408a5424ba824c55ca2a3a2a4b7c7156dcfac3fe66eafa23073a4b6d25637b8587926b41159c912e49dbf8143dd32084885d84f82bdd615a4582a54916fb3c1581418610194809502400c9592c6eeb0f9cb87244ae8cfacb83807391eec72d5f2dd7de0082dd2d4364d8d5a449b09f50573c661930049961684474b9665a202b52c1c66cba861f5afe2e4c09dd419ccb83a048af1f0eca750728b6b5d8d8a2e94e68fc05977128e31b6f99008c4d7c3ef01ba1d9df04581db492fd9cf2507c3383a6581c0ab2f937540034aac9f0becbdfcbde25dc0f05bde34f9a512d54d43c846513a1b3306553684180865bd6722c9467204cc5362a0de64100e268342a14cd884bc507135ba3f81eafc8dfa823c88055121a8a9548746d0b9edaaa6e6518b229d14c06fb198751a853f12a4ec55dda54f89b588b04740836614b305f804864128547057b283c9cafc22c8a49a589ba3584c8f5b58d7735efd9cb5542fd3b7abccb36a2fdb879465c6fc1f5a289e1cf08913103bc4db14bcc449cca4494a8d2f317ec5f5accb290cd54c274cc977aa1d21461453c438e96e62f847f439758283585421928090c56ad640026e374493ba17d5943fb507095aa69536d28d3f0a2655b956575ead1537b42914b58d7d57d76a608d770e377603f5361e039ecfc279100e522443f540053c07c0ee0ab15108ec22d23bff1d4f8778e0548a6995e0a5fce24c6626b86d5187ae6ae40db20c8fe6d42237cd3ad0cd205fc5daa2e9967115886c6874ca483fc96a900d1a01cd6b695322dec1daa1a4f86ba1192b28b0f711170999651cc85f0116e2ed7351e52a7d5959b8984245487bd245982a9b793955f5459d953a909d93cd24f57ad4671b62b348529cacc5cb67035d4161a4ac9dfb6958dd05beb3dac5ea2af809b30581083edc72555ad78b5674565b93b56b0a3ee30cad815519b5c55e9e4c88365315aa385f9568225e86cea9505c81ecb784903153080aa38cbed46330755fe60a21bd83d8ca5fb4d2cd7cf1efc763b835db84acf3c32a6964874f21446cc1868e24060163b191b8869366f52f69b6fdf30333037a4b3e14b49c750cdc145a73901173c09308c857e42ab506c45c50847281d6420c1e6d8b7344b547472b975ac72d90115b406512aa42f7ef5f08e1339e5dd2e64d81a36bd37b64cf77247cd4a6a1b1f90825388f3790c836b06208726200a246a0e10d60c68f6142c7b83b555aed3e0e63d8fe7f8dc471b71560bf82a8277d120890665c99c796c7cd6afe80d06db78274d99fc195e7a648bde30582c34dd3d542c6ea3ba4a3a0d8f0dbb41717a0cc366f73c0aa01337c93e6090de61838e2ab0755eab8a1fd0848abe31a651b3c6d8136c08dd43b367ad759a693e76a8f59b34ccd8bed724f936775e586ccf28f4d43f3ffbfe994c19fb6695386d0ec5f02211a964dd1d55f298986f888f9d56ca693c95247ccd545235086321681692c6de8ce6ad778e2087519384b373d39d1631d420d2f5b29392783daf0c2b1935a1c0a9954ae39aa4d93e80e89a6da3e6b994d8354bf40f9c1a7a092186040fbf6637c2f7f9cde4f597e77612ba121e94eeac39697aa0b1314d63b83b6850d1e435f4d0db1be4570f6c90aea91191f8de78fcfab537e5218fc1cdaa59533f7f1729016d4bd859bcbfbeff0e9f0342a376cc35eefb0a22b72d0deb3033d924876767ad87cf515b91b758ada17af69d926ec2ee8e34dafa815f53bf65450d2de94e35e1e42a3d6def9bc61b4f684ac8e0a07e714f3b7347c22f6d287cce5eb149773cdc1ffb357aa3a1c6db1483a2e7cfd9e47cf7107a4fe61e20f62e38bdb3a371701b983688e193cb5dfccfe4f2f9799f1d7d4909e0269c5395f2d0d3d97219b6e2a81c6973b703e8c704edee48d9c66f002a5f84aa7a5ede5a3e2820b3a5b7652962f0f0581eba932cb03b443fd9a18bc1baabb27d88f69e04f31f0afc0176f3fa6a7ce5a9dbb8cc58a8ad05136674bb84181f61d8ad42cd314d9ba609f7d3a32596e415ece0527a1afed564bc4b17068f9dfa304bb50a84bd1cc372c4a11ea01b1c7b06301f6edc51e0d79387a8b3a43d972f3e65b8f24193272ff57931e596dae5b729e5268a119317e282cd217ac4ac8d740180ff2e123582a1fa8aa09eb1589d46ec1bd4d7b5389318153815370d957fe399a949579609258f64492280ff543e5f8fd99afec3e6836c5d1ed5b541729ace4e79b4ec16be43f6e289c1e2b7a56d35a82a7e8cf93aac6805d6d7fcdb48ec6a19587a6b49318051769485a95c8679222518d05c80f8c905fbc55cf0c16ec0a484c4ab2c10be8c78887378d0523117c9884db3142d897116499a669caa74b15dfad3ce863579b286d5999b9d45a85540d6237d51723ee23dbece2fe03d5bacbc87179c6ba272d5caeb9798a856d1bd5565ce010df68debcb3c0cd7273deea0475412e2de9300ddead6b8d43d2ae19a2dd52d8ed2c0be04f7174ca13d8e98b8985fb03bf65fb6867fbf9fb4b5a3b6005ca7b6227ed71a468850f8b61783db32ca22544cc3999e7ce80369933456759078c1a95ed95adfaea957c627f53e32f5cab8a15ee9cea68a542bf0f98e877e86c9dc1a1399b250d5945bf7bc2725edaa24e55e7f05ac9b3af9d1c02f41276f4f2a38900a0066970a74366d79bd6c4a5352a88193877b50cd9176ea11fed4ad4a1f37388dafdb3325740754329377228000125e5d497b7c9b2b92f2e6c7d7e59fb7a06581764849ff34936514b89d42f3f77565897c9fd765a76f4c92c55460725862040f3af74bd633d4db64c0ecc52fafaeaf4a3a8cdc126a9ee2e6afce3c963733366628088a942dcd56b0fc0106ff2df26925dc6d300bd52d5b729d167b1721e6e86c0026d86c907f90ec89cae2c0cb961b650fd46d6ca4a747b02d55e9e9b68a1910d8971a731b43a5b56bafeae560d13758e1117761ba66c87343f182355c92fe510ca69227544c036db9aff944caa740ca7ceeadf072253f47f748e7d6e0a560718c147df7ee47dd4567ccb6a5626fb723a7d926eef5b9f60f0bcc96c362357b31d5a5eae4d54f065431835a1455a5904dbb2c89c3deee6f333c6ea59b5457246632d118c7c61e2e15591170b10e602ea214ca94a05dc01a179152f2fde24e6acc6d8dc58889500bbba0a769cda9b12a700a8b77088b7b72b205872188275bcbb02621fbcabdc78f94687568754fe9d1d08d171ca09ce90a45b4c04d20b65c4dad0dc5e3f5abc4943c621641df76950fb4433bdfb202b20a24307ac6c17d15383f8d9b1ea111a299e6d76dad26b57d66b787e79557207d855c182ec18ccbf0401250ed33e88ff63c4307a8e6e4e2a78c875eaa5a734e2a02c5dacd7b21f62813cbcbc6fab5597eabe40468b29d57e3ca1619dbb835e3e879b24f3c6c708fdaf8eaf258b423a8c75b063c933d73063eb931879e7a805b1fe4cd69d73da11b11ba39694f5ebd72eb705e1d7172a9d5a66e7818cf74bb766b47dd67df98fdd915d329adace6f1ed1943ba9e2a5271f5a600bed29a954d1d39eea143f76c5ef3d4a907a9c9025a1ead1a0195289265b8bc574d0598eba41fc071ede9fbb5677a9011bc56b9510b05749cf7455dacec10bcedc94a79444caeedd9736b60994c3dbe6aeda4afacfe20583c91e902da237dc6a11f815012fdbf391f653dc285f772a2a73a901e9971573e143b9a33b81b8ede7a3f6657724d56ee3df688aa2f646b01e73f5284810d91e12dc0c894dd2e148cdcbaf1cb27226a5e57258c02280ce3eeee05d2bd70221cfa80ea49be496f553f32fa82bd4786e1e4c92d4f02e35a9f15cffa726410be8e3c9986f5c6347313c2f07b0943ef38b5b6eb32dac74f296cf206485aa726363099ca9827eb062499ceb7dddbdb335ceac371113471a182278ed9a417b459a8542b13b5e2b08c0f1ae8b0cabef689792b83422f4ec0891e98c49d8f6d9f8a3586171486e02b9fbccd51b3fb7099471b4c6e2e14489d400c11091e0f0b204c7efa8b17fa5392bebc639fb3f5f98b17cf39aa98bb9354116419b7faca0aca63f68add0d0bc74c77797fbde55ea060abfb50095be307470f18aab909659fbc3986ee642e4244cdc7971b40f15514d192c7b3c1841abd01968181c1b342a5bfe3ef4f34db1c8297db219e9323eecb0b43d496edc3616d7afa5d7ebc1c25a3237fd60fcc9ffbc2ac2f41dc629639417b69d6e2dd443bafe3ba662fb358cb798c27ba720d6185cdb33bafc4f1331ea94c331ece95b1e65fcff2905f9af3aca76ba6213e4de0895fcf9e9b36a0e7ec9ffac1b86a8343247404440eca8c921ed974cf5e1051f39d298326e1015ae3756c0ccb974eb4497b27a48a13869d8ba97286c9062b91e0512e668f65313b6c8f949f2bdcca32552bda35517cfba172b8d969f1640b658316f36d611585dadd6178b36f5f58e5898e5e32a74094f79641aec313211e37217a93beab2c69e57f97666ff2c0f34f221c56db534ca9aaae279d14bc9d82fb52c1a54a02914c95baf156636fc501f5e2308386b63fb40e82289e65f4e0b17ab852c4fcf4811a5855f91a5fff78f32e7ff497c9f0a0c31cd2b06d5deff2a31bcce226e231b8c6fad90a436bfc804f6db35486caddf724860645aadaa9df722f0e6312ea6ae22522e23246d762cf8888571089761d93490ec11666c9877fbec5f5cb9939b99a8e6fd0789efb91d867e43dc8394ef42677b3072006fe76d56998f6567eea4f8d494f0012d3ae8d00d04fccecca74775f94773150c07c71ee67cc04cd8fcee9f834893d7f184fc3cbadabde553d150c2d2a1b212482743208eceaa911c73466409bf1342dcb32f8e90bb912ad410c65a1634986e364b30c56332b8c52cca3c41c6408ef3a28ffb162d50982133c88c7e46c8d51813d75c852cce3f9692275f8dee3a7eb6c2e4ad30ccb938b0ae332af67abb1cb3aef2fe7c436155bbae3215243d554eb480dbe1bec762e5a531206687623d6af4cc2ce92cba43832d940a9a143c1c414e0a9f9ca543e07748339a0b57388ab49f507273154d71dcee4b2d850a6de9803ec7469c3f6ef0446226ddc3a4f69bea5b2c4c726a4c7446a8b506eabacb4aef381a8896eeb4f2584c014722c557b56ac9a2a6b4a34cc9cfa5e7c47891fe2aa963846c45362f2cf51e14e38e92f2873b66b23dc831ce33bf96dc43e5ee17fe3cbcd9fb5701ce3bbe3c9bcf6b3169342eced5c19e9edf80765d63f4a666f3636116001cc7530198bc71ae052832d4b07b299ba2f6fa6120faff771bb9cc5202e876225c2e74878e2fa1744f8f166c25f1d83f0577b101e347b0f07a823579eb68134883f60d22f90b3d626a99a81144519161589b0b343136b26dddb9b9b03310a1b710ad03d153815ee69c0a79b448fd37384b1531774b09017a6ad78bbde07e733eda67c1eed95996745da59e9d25c5fa988143c5dee61cff508a1247e6a728e69cfa02efca0d73ebe8aded4b567d5b480418b4d6d5f33daad83b9733c5eb3d673c57edaab111b97df9c19b9ab767dadac523316f91a4f721ad132037e8505f3e7bbeabd1eb1bff46f869d0cdac7fb748e0a4d04183a597aebe0cb06a234a42e4ebac4689af64fa67419e895738bcf399e293edcf66a616947a5b68f46be93527f009d22c70fe2641da69a2d039e5a480a27eb25a4a741b64b2f28be7585cf990f4cdb43fc78d1aaa6c99269b6ccf920218588836a94546b056578da0c3a694e0830187491f88b7e12031baf876dd5c70fd3f68a44258609b5f9bb76d747d9c85ea58adb7ad2a6f598c7ace9103f5275189b496b3663ebc1c0acac6bbf8eae6831e23362d5bfad89500ff8d96f9ffd1ff9cdc38f011bffe5282f43e495b402706170706c69636174696f6e2f6a736f6e03676465666c617465";
 
   const dataDecoded = decodeRainMetaDocument(rainDocumentEncoded);
