+++ utils/orderBook/index.ts
@@ -3,8 +3,14 @@ import { ethers  } from "hardhat";
 import { BigNumber } from "ethers";
 import { eighteenZeros, ONE } from "../constants";
 import { getEventArgs } from "../events"; 
-import { fixedPointMul } from "../math";
-import { compareStructs } from "../test";
+import { fixedPointDiv, fixedPointMul } from "../math";
+import { compareStructs } from "../test"; 
+
+export const scaleOutputMax = async (orderRatio: string, decimals : number) => { 
+    let a = ethers.BigNumber.from(1 + "0".repeat(decimals)).mul(1000) 
+    let b = ethers.BigNumber.from(orderRatio) 
+    return fixedPointDiv(a,b) 
+}
 
 // Hacky Util
 export const prbScale = async (index: number, orderRatio: string) => {  
@@ -30,17 +36,15 @@ export const prbScale = async (index: number, orderRatio: string) => {
   // Hacky Util 
   export const scaleRatio = async(ratio: BigNumber, aDecimals: number,bDecimals: number) => {   
    
-    let maxRatio
-    maxRatio = fixedPointMul(
-      ratio,
-      ethers.BigNumber.from(10).pow(18 + aDecimals - bDecimals)
-    )  
-    if(maxRatio.mod(10).gt(5)){ 
-      maxRatio = maxRatio.add(1)  
-    }    
+    const decimalDiff = ethers.BigNumber.from(10).pow(18 + aDecimals - bDecimals)
+    let maxRatio =  ratio.mul(decimalDiff).div(ONE)
+
+    if(ratio.mul(decimalDiff).mod(ONE).gt(ethers.BigNumber.from('1'))){
+      maxRatio = maxRatio.add(1)
+    } 
     
     return maxRatio
-  }
+  } 
 
 // Util to take order
 export const takeOrder = async (
@@ -55,8 +59,11 @@ export const takeOrder = async (
     orderRatio // ratio to scale by index
     ) => {
    
+    // Scaling ratio as batch index increases 
+    const ratio = await prbScale(index,orderRatio)  
+
     // Deposit max amount per batch
-    const amountB = ethers.BigNumber.from("1000" + eighteenZeros);
+    const amountB = await scaleOutputMax(ratio.toString(),18) ;
 
     const depositConfigStructAlice = {
       token: tokenB.address,
@@ -79,8 +86,7 @@ export const takeOrder = async (
      signedContext : []
     }; 
 
-    // Scaling ratio as batch index increases 
-    const ratio = await prbScale(index,orderRatio) 
+    
 
     const takeOrdersConfigStruct = {
       output: tokenA.address,
@@ -92,7 +98,11 @@ export const takeOrder = async (
     };
    
 
-    const amountA = amountB.mul(ratio).div(ONE)
+    // Tracking Input Token amount 1000 
+    let amountA = amountB.mul(ratio).div(ONE); 
+    if(amountB.mul(ratio).mod(ONE).gt(ethers.BigNumber.from('1'))){
+      amountA = amountA.add(1)
+    } 
 
     await tokenA.transfer(taker.address, amountA);
     await tokenA.connect(taker).approve(orderBook.address, amountA); 
@@ -114,4 +124,5 @@ export const takeOrder = async (
     assert(output.eq(amountA), "wrong output");
 
     compareStructs(config, takeOrderConfigStruct); 
-}
+}  
+
